# upravljanjevremenom/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Time Manager API",
        "description": "REST API sistem za upravljanje vremenom. \nOmogućava korisnicima upravljanje zadacima, beleškama i obaveštenjima, \npraćenje statusa zadataka, kao i pregled statistike rada korisnika. \nSistem podržava uloge (admin, regular i premium korisnik).",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/obavestenja": {
            "get": {
                "tags": [
                    "Obaveštenja"
                ],
                "summary": "Lista svih obaveštenja",
                "operationId": "55e043b458ea4da569fd85bd9a4f2f28",
                "responses": {
                    "200": {
                        "description": "Uspešno vraćena lista obaveštenja"
                    }
                }
            },
            "post": {
                "tags": [
                    "Obaveštenja"
                ],
                "summary": "Kreiranje novog obaveštenja",
                "operationId": "a411c94de7fba1df4d6b180dd7713723",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "poruka",
                                    "poslato",
                                    "nacin_slanja",
                                    "korisnik_id"
                                ],
                                "properties": {
                                    "poruka": {
                                        "type": "string",
                                        "example": "Podsetnik za sastanak u 15h"
                                    },
                                    "poslato": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-02-23 14:00:00"
                                    },
                                    "nacin_slanja": {
                                        "type": "string",
                                        "example": "email"
                                    },
                                    "korisnik_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Obaveštenje uspešno kreirano"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                }
            }
        },
        "/api/obavestenja/{id}": {
            "get": {
                "tags": [
                    "Obaveštenja"
                ],
                "summary": "Prikaz jednog obaveštenja",
                "operationId": "2916ad56517d91b984ab5bbd22f7e05b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Obaveštenje pronađeno"
                    },
                    "404": {
                        "description": "Obaveštenje nije pronađeno"
                    }
                }
            },
            "put": {
                "tags": [
                    "Obaveštenja"
                ],
                "summary": "Izmena obaveštenja",
                "operationId": "6c6b69daf9aef6d572ba2b6c7a488af5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "poruka": {
                                        "type": "string",
                                        "example": "Izmenjena poruka"
                                    },
                                    "poslato": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-02-23 15:00:00"
                                    },
                                    "nacin_slanja": {
                                        "type": "string",
                                        "example": "sms"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Obaveštenje uspešno izmenjeno"
                    },
                    "404": {
                        "description": "Obaveštenje nije pronađeno"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Obaveštenja"
                ],
                "summary": "Brisanje obaveštenja",
                "operationId": "26265e3d69a0f23430f8033f1c27e919",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Obaveštenje uspešno obrisano"
                    },
                    "404": {
                        "description": "Obaveštenje nije pronađeno"
                    }
                }
            }
        },
        "/api/podsetnici": {
            "get": {
                "tags": [
                    "Podsetnici"
                ],
                "summary": "Lista svih podsetnika",
                "operationId": "fb5d2c15cc2755e0ade3f3fe9194312c",
                "responses": {
                    "200": {
                        "description": "Uspešno vraćena lista podsetnika"
                    }
                }
            },
            "post": {
                "tags": [
                    "Podsetnici"
                ],
                "summary": "Kreiranje novog podsetnika",
                "operationId": "3ded20b0a0ed5bd7d45a5dd61afdd958",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vreme",
                                    "aktivan",
                                    "korisnik_id"
                                ],
                                "properties": {
                                    "vreme": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-02-23 15:00:00"
                                    },
                                    "aktivan": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "korisnik_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Podsetnik uspešno kreiran"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                }
            }
        },
        "/api/podsetnici/{id}": {
            "get": {
                "tags": [
                    "Podsetnici"
                ],
                "summary": "Prikaz jednog podsetnika",
                "operationId": "e7753a6b98ff1e821e583eaec3d5a041",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Podsetnik pronađen"
                    },
                    "404": {
                        "description": "Podsetnik nije pronađen"
                    }
                }
            },
            "put": {
                "tags": [
                    "Podsetnici"
                ],
                "summary": "Izmena podsetnika",
                "operationId": "1bdc85e44dfeef728f6a0dc3642a8f96",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "vreme": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-02-23 18:00:00"
                                    },
                                    "aktivan": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Podsetnik uspešno izmenjen"
                    },
                    "404": {
                        "description": "Podsetnik nije pronađen"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Podsetnici"
                ],
                "summary": "Brisanje podsetnika",
                "operationId": "5bc195ac857ed7aa6f8b3f672cb3001b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Podsetnik uspešno obrisan"
                    },
                    "404": {
                        "description": "Podsetnik nije pronađen"
                    }
                }
            }
        },
...
    },
...
}